Reworked event loading
