--- src/events/emojiCreate.ts
@@ -1,9 +1,0 @@
-import Event from "../core/event";
-import {updateGlobalEmoteRegistry} from "../core/libd";
-
-export default new Event<"emojiCreate">({
-    on(emote) {
-        console.log(`Updated emote registry. ${emote.name}`);
-        updateGlobalEmoteRegistry();
-    }
-});
