--- src/events/guildDelete.ts
@@ -1,9 +1,0 @@
-import Event from "../core/event";
-import {updateGlobalEmoteRegistry} from "../core/libd";
-
-export default new Event<"guildDelete">({
-    on() {
-        console.log("Updated emote registry.");
-        updateGlobalEmoteRegistry();
-    }
-});
